RSC: Refactor build process
